Feat: creating container docker and fixing several bugs in all routers for do it works
