+++ huoc-front/Dockerfile
@@ -1,0 +1,23 @@
+# Etapa 1: build
+FROM node:20-alpine AS builder
+
+WORKDIR /app
+
+COPY package*.json ./
+RUN npm install
+
+COPY . .
+RUN npm run build
+
+# Etapa 2: produção
+FROM node:20-alpine AS runner
+
+WORKDIR /app
+
+COPY --from=builder /app/public ./public
+COPY --from=builder /app/.next ./.next
+COPY --from=builder /app/node_modules ./node_modules
+COPY --from=builder /app/package.json ./package.json
+
+# As variáveis virão do docker-compose.yml via `env_file`
+CMD ["npm", "start"]
