+++ huoc-front/docker-compose.yml
@@ -1,0 +1,7 @@
+services:
+  huoc-front:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    ports:
+      - "3000:3000"
